requirements.txt
mecab-python3==0.7
sanic==0.5.4
sanic-cors==0.6.0.2
